chore: adopta baseline fiscal-download de Robert (STM-719 uuid-key + addenda NC desde portal) como version final

+++ APP03022-mrch.batch.somx.fiscal-download/src/main/java/com/sodimac/batch/fiscal/download/repository/sap/ConceptoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConceptoRepository extends JpaRepository<ConceptoEntity, Integer> {
+public interface ConceptoRepository extends JpaRepository<ConceptoEntity, Long> {
 }
